Deploy v1.4.0 to production

+++ src/courses/ui/edit-config/schema/edit-course-config-action-schema.ts
@@ -8,9 +8,9 @@ import { default_maximum_interval } from "ts-fsrs";
 export const EditCourseConfigActionSchema = z.object({
   enrollmentId: ObjectIdSchema,
   enableFuzz: z.boolean(),
-  maximumInterval: z.number().min(1).max(default_maximum_interval),
+  maximumInterval: z.number().int().min(1).max(default_maximum_interval),
   requestRetention: z.number().min(0).max(1),
-  dailyNewCardsCount: z.number().min(1).max(100),
+  dailyNewCardsCount: z.number().int().min(1).max(100),
   showAdvancedRatingOptions: z.boolean(),
 });
