+++ src/courses/ui/profile-courses/schemas/paginate-courses-by-author-action-schema.ts
@@ -6,7 +6,7 @@ import { z } from "@/i18n/zod";
 export const PaginateCoursesByAuthorActionSchema = z.object({
   profileId: z.string(),
   paginationToken: z.string().optional(),
-  limit: z.number().optional(),
+  limit: z.number().int().optional(),
 });
 
 /**
